add options to use new ZMQ Context on for every ZStack (useful in tests)

+++ stp_zmq/zstack.py
@@ -44,7 +44,6 @@ from stp_core.validators.message_length_validator import MessageLenValidator
 
 logger = getlogger()
 
-
 Quota = NamedTuple("Quota", [("count", int), ("size", int)])
 
 
@@ -238,7 +237,7 @@ class ZStack(NetworkInterface):
     @staticmethod
     def keyDirNames():
         return ZStack.PublicKeyDirName, ZStack.PrivateKeyDirName, \
-            ZStack.VerifKeyDirName, ZStack.SigKeyDirName
+               ZStack.VerifKeyDirName, ZStack.SigKeyDirName
 
     @staticmethod
     def getHaFromLocal(name, basedirpath):
@@ -343,7 +342,10 @@ class ZStack(NetworkInterface):
 
     def start(self, restricted=None, reSetupAuth=True):
         # self.ctx = test.asyncio.Context.instance()
-        self.ctx = zmq.Context() #zmq.Context.instance()
+        if self.config.NEW_CTXT_INSTANCE:
+            self.ctx = zmq.Context()
+        else:
+            self.ctx = zmq.Context()
         if self.config.MAX_SOCKETS:
             self.ctx.MAX_SOCKETS = self.config.MAX_SOCKETS
         restricted = self.restricted if restricted is None else restricted
@@ -357,7 +359,6 @@ class ZStack(NetworkInterface):
         if self.opened:
             logger.display('stack {} closing its listener'.format(self), extra={"cli": False, "demo": False})
             self.close()
-        #self.teardownAuth()
         logger.display("stack {} stopped".format(self), extra={"cli": False, "demo": False})
 
     @property
@@ -405,31 +406,34 @@ class ZStack(NetworkInterface):
                 time.sleep(sleep_between_bind_retries)
 
     def close(self):
-        self.ctx.destroy(linger=0)
-        self.listener = None
-        self._remotes = {}
-        self.remotesByKeys = {}
-        self._conns = set()
+        if self.config.NEW_CTXT_INSTANCE:
+            self.ctx.destroy(linger=0)
+            self.listener = None
+            self._remotes = {}
+            self.remotesByKeys = {}
+            self._conns = set()
+        else:
+            if self.listener_monitor is not None:
+                self.listener.disable_monitor()
+                self.listener_monitor = None
+            self.listener.unbind(self.listener.LAST_ENDPOINT)
+            self.listener.close(linger=0)
+            self.listener = None
+            logger.debug('{} starting to disconnect remotes'.format(self))
+            for r in self.remotes.values():
+                r.disconnect()
+                self.remotesByKeys.pop(r.publicKey, None)
 
-        # if self.listener_monitor is not None:
-        #     self.listener.disable_monitor()
-        #     self.listener_monitor = None
-        # self.listener.unbind(self.listener.LAST_ENDPOINT)
-        # self.listener.close(linger=0)
-        # self.listener = None
-        # logger.debug('{} starting to disconnect remotes'.format(self))
-        # for r in self.remotes.values():
-        #     r.disconnect()
-        #     self.remotesByKeys.pop(r.publicKey, None)
-        #
-        # self._remotes = {}
-        # if self.remotesByKeys:
-        #     logger.debug('{} found remotes that were only in remotesByKeys and '
-        #                  'not in remotes. This is suspicious')
-        #     for r in self.remotesByKeys.values():
-        #         r.disconnect()
-        #     self.remotesByKeys = {}
-        # self._conns = set()
+            self._remotes = {}
+            if self.remotesByKeys:
+                logger.debug('{} found remotes that were only in remotesByKeys and '
+                             'not in remotes. This is suspicious')
+                for r in self.remotesByKeys.values():
+                    r.disconnect()
+                self.remotesByKeys = {}
+            self._conns = set()
+
+            self.teardownAuth()
 
     @property
     def selfEncKeys(self):
@@ -560,7 +564,8 @@ class ZStack(NetworkInterface):
                 except zmq.Again as e:
                     break
                 except zmq.ZMQError as e:
-                    logger.debug("Strange ZMQ behaviour during node-to-node message receiving, experienced {}".format(e))
+                    logger.debug(
+                        "Strange ZMQ behaviour during node-to-node message receiving, experienced {}".format(e))
             if i > 0:
                 logger.trace('{} got {} messages through remote {}'.
                              format(self, i, remote))
@@ -573,8 +578,8 @@ class ZStack(NetworkInterface):
         # These checks are kept here and not moved to a function since
         # `_serviceStack` is called very often and function call is an overhead
         if self.config.ENABLE_HEARTBEATS and (
-            self.last_heartbeat_at is None or
-            (time.perf_counter() - self.last_heartbeat_at) >=
+                self.last_heartbeat_at is None or
+                (time.perf_counter() - self.last_heartbeat_at) >=
                 self.config.HEARTBEAT_FREQ):
             self.send_heartbeats()
 
@@ -597,7 +602,7 @@ class ZStack(NetworkInterface):
             frm = self.remotesByKeys[ident].name \
                 if ident in self.remotesByKeys else ident
 
-            if ident in self.remotesByKeys:
+            if not self.config.RETRY_CONNECT and ident in self.remotesByKeys:
                 self.remotesByKeys[ident].setConnected()
 
             if self.handlePingPong(msg, frm, ident):
